requirements.txt
streamlit==1.22.0
twilio==8.0.0
python-dotenv==0.21.0
pandas